Adding requirements on auth*
This includes a section rename. I think this needs to eventually be covered
within the working group, as I foresee analysis of the security and trust models
that may arise from later work describe this as something that is required - in
existing protocols we have fairly primitive things in place - RTMP has access
keys, HTTP based protocols use a wide variety of methods.