Rephrasing "configurable latency"
I think a better way of proposing this issue is one of a trade-off between
guarantee of delivery, and of delay as receivers may drop stuff and not use
media if it either never arrives or because it arrived too late. This kind of
phrasing should also clear up a discussion around "stream vs datagram" because
the requirements defined here wil largely determine what types are used -
streams for reliable in order, datagram for lossy out-of-order.