Requirements Requirements Requirements (#90)
* Cleaning up encryption requirements

* Adding requirement on contribution and distribution

* Adding requirements on auth*

This includes a section rename. I think this needs to eventually be covered
within the working group, as I foresee analysis of the security and trust models
that may arise from later work describe this as something that is required - in
existing protocols we have fairly primitive things in place - RTMP has access
keys, HTTP based protocols use a wide variety of methods.

* Adding bit about encapsulation

I'm not sure consensus has been finalised about which means of encapsulation
should be required; for now saying that if the WG is going to use an existing
format, we should have agility within the protocol.

* Rephrasing "configurable latency"

I think a better way of proposing this issue is one of a trade-off between
guarantee of delivery, and of delay as receivers may drop stuff and not use
media if it either never arrives or because it arrived too late. This kind of
phrasing should also clear up a discussion around "stream vs datagram" because
the requirements defined here wil largely determine what types are used -
streams for reliable in order, datagram for lossy out-of-order.

* Adding requirements on WebTransport

* Adding bit on codec negotiation and agility

This has been moved up since it applies to both producing and receiving.

* Adding requirements on negotiation and variants

* Updating text on publishing

* Fixing format error

* Fixing nits (thanks Spencer!)

* Further changes based on Spencer's feedback

* Removing the billing relationship requirement

This really can be addressed via authentication, and through the data model used
to represent media. This is further complicated by providers which may have
differing sales models.

* Covering data model as a concept

Put as a higher level concept, but ties together hopefully a lot of other
discussions.

* Fixing various things based on feedback (thanks Spencer!)

* Updating bit on WebTransport

This gets around the word "raw QUIC" with hopefully some better language.

* This should clear up negotiation